requirements.txt
pystray~=0.19.4
psutil~=5.9.5
Pillow~=9.5.0